// site/src/types/SystemState.ts
export enum SystemState {
    INITIALIZING = "INITIALIZING",
    IDLE = "IDLE",
    STARTING = "STARTING",
    RUNNING = "RUNNING",
    PAUSING = "PAUSING",
    PAUSED = "PAUSED",
    UNPAUSING = "UNPAUSING",
    STOPPING = "STOPPING",
    COMPLETED = "COMPLETED",
    ERROR = "ERROR",
    REBOOTING = "REBOOTING",
}